moving the submit function to an external file
